Update to SDK 4.2 : New video player and jog control
git-svn-id: https://sbgracing.sourcerepo.com/sbgracing/RacePad/RacePad@113 9925204c-63e1-4f11-ae18-372a25b3a23a

+++ Classes/UIViews/PitWindowView.m
@@ -77,7 +77,7 @@
 	float currentUserScale = userScale;
 	float centreX = x / current_size_.width ;
 		
-	if(fabs(currentUserScale) < 0.001 || fabs(scale) < 0.001)
+	if(fabsf(currentUserScale) < 0.001 || fabsf(scale) < 0.001)
 		return;
 		
 	// Calculate where the centre point is in the untransformed window
